Finishing the wizard

+++ app/src/main/java/com/vbshkn/ikollect/data/local/entity/AlbumArtistCrossRef.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["albumId", "artistId"])
 data class AlbumArtistCrossRef(
-    val albumId: Int,
+    val albumId: Long,
     val artistId: Long
 )
