+++ app/src/main/java/com/vbshkn/ikollect/data/remote/dao/SearchResponse.kt
@@ -9,9 +9,9 @@ data class SearchResponse(
 
 data class SearchResult(
     @SerializedName("id")
-    val id: Int,
+    val id: Long,
     @SerializedName("master_id")
-    val masterId: Int,
+    val masterId: Long,
     @SerializedName("title")
     val title: String, // Artist - Release
     @SerializedName("year")
